refactor: ContentItem
addresses https://github.com/datacite/akita/issues/6
